Move kptfile util functions to public
