+++ internal/util/man/man.go
@@ -24,8 +24,8 @@ import (
 	"strings"
 
 	"github.com/cpuguy83/go-md2man/v2/md2man"
-	"github.com/kptdev/kpt/internal/pkg"
 	v1 "github.com/kptdev/kpt/pkg/api/kptfile/v1"
+	"github.com/kptdev/kpt/pkg/kptfile/kptfileutil"
 	"sigs.k8s.io/kustomize/kyaml/errors"
 	"sigs.k8s.io/kustomize/kyaml/filesys"
 )
@@ -57,7 +57,7 @@ func (m Command) Run() error {
 	}
 
 	// lookup the path to the man page
-	k, err := pkg.ReadKptfile(filesys.FileSystemOrOnDisk{}, m.Path)
+	k, err := kptfileutil.ReadKptfile(filesys.FileSystemOrOnDisk{}, m.Path)
 	if err != nil {
 		return err
 	}
