+++ internal/util/pkgutil/pkgutil.go
@@ -327,7 +327,7 @@ func RoundTripKptfilesInPkg(pkgPath string) error {
 	pkgsPaths = append(pkgsPaths, pkgPath)
 
 	for _, pkgPath := range pkgsPaths {
-		kf, err := pkg.ReadKptfile(filesys.FileSystemOrOnDisk{}, pkgPath)
+		kf, err := kptfileutil.ReadKptfile(filesys.FileSystemOrOnDisk{}, pkgPath)
 		if err != nil {
 			// do not throw error if formatting fails
 			return err
